SPCI ALP2+: Initialise OP-TEE with SPCI messaging framework
This patch uses the OP-TEE resource description to initialise it using the
messaging framework described in SPCI Alpha 2. Earlier patches add support for
specifying the RD as the TOS FW configuration. The RD is a device tree with
nodes that describe the memory regions OP-TEE needs to operate correctly.

The SPM obtains the RD from platform code and converts it into a SPCI
architected initialisation message. This message describes the memory region
required by OP-TEE and the secure and non-secure RX/TX buffers allocated by the
SPM on its behalf. Memory for the buffers was reserved in earlier patches. The
initialisation message is not described in the SPCI Alpha 2 specification. It
will be added in the next spec version.

OP-TEE issues SPCI_MSG_RECV to indicate successful initialisation. The SPCI
Alpha 2 SMCs fall under the standard service space in the SMCCCv1.1. Hence, this
patch chooses the right SMCCC version support. Furthermore, it ensures that the
right SPM SMC handler is invoked when OP-TEE returns control.

Signed-off-by: Achin Gupta <achin.gupta@arm.com>