SPCI ALP2+: Implement SPCI_MSG_SEND suitable for normal world
This patch implements the SPCI_MSG_SEND interface to enable the normal world to
post a message for the single SP in the secure world. The client and SP ids are
assumed to be 0. They are distinguished by the caller's security state. A helper
function has been added that returns either the RX buffer of the specified
target or the TX buffer of the specified source.

Signed-off-by: Achin Gupta <achin.gupta@arm.com>