SPCI ALP2+: Implement SPCI_RUN suitable for a Normal world OS
This patch implements the SPCI_RUN interface which is used by the OS to allocate
CPU cycles to the SP. It is assumed that the caller used scheduling model B as
described in the SPCI spec. Further changes to this implementation might be
required to support both scheduling models described in the spec.

Signed-off-by: Achin Gupta <achin.gupta@arm.com>