SPCI ALP2+: Implement SPCI_MSG_SEND_RECV and completion of SPCI_RUN
This patch implements this interface so that it can be invoked by the OP-TEE
SP. This is a deviation from the specification as this interface must be invoked
only by S-EL0 SPs. However, the semantics of the interface are more suitable for
preserving the existing run time model of OP-TEE with its Linux driver. These
changes will be added to the next version of the spec.

This patch also demonstrates how an invocation of the SPCI_RUN interface from
the normal world is completed in response to a SPCI_MSG_SEND_RECV from the
secure world. Support for all completion reasons has not been added as yet.

Signed-off-by: Achin Gupta <achin.gupta@arm.com>